requirements.txt
fastapi
uvicorn
fastmcp
requests
openai
python-dotenv